requirements.txt
python-pptx
fastapi
uvicorn
diagrams
pylint